requirements.txt
Django==4.2.10
Pillow==10.1.0
django-unfold==0.57.0